update: split room detail page into components
