+++ src/components/room/ReserveInfo.jsx
@@ -1,0 +1,25 @@
+const ReserveInfo = () => {
+  return (
+    <ul className="space-y-3">
+      {[
+        '入住時間為下午3點，退房時間為上午11點',
+        '如需延遲退房，請提前與櫃檯人員聯繫，視當日房況為主（可能會產生額外費用）',
+        '若您有任何疑問或需要協助，請隨時與我們的櫃檯人員聯繫，我們將竭誠為您服務',
+        '為了確保所有房客的安全和舒適，我們提醒您遵守飯店的相關規定和禮儀',
+        '請勿在房間內抽煙，若有抽煙需求，可以使用設在酒店各樓層的專用吸煙區',
+        '若您需要額外的毛巾、盥洗用品或其他物品，請聯繫櫃檯，我們將盡快為您提供',
+        '請愛惜房內的設施和物品，如有損壞或遺失，將依照飯店相關規定進行賠償',
+        '如您在入住期間遇到任何問題或有特殊需求，請隨時向櫃檯人員反映，我們將盡力協助解決',
+        '為了維護其他房客的權益和安寧，請避免在深夜時段大聲喧嘩或播放音樂',
+        '為了維護房間的最佳狀態，我們禁止在房間內烹飪或使用加熱設備，感謝您的理解與配合',
+      ].map((text, i) => (
+        <li key={i} className="flex">
+          <span className="w-6 text-right">{i + 1}.</span>
+          <span className="text-sm text-[#4B4B4B] leading-7">{text}</span>
+        </li>
+      ))}
+    </ul>
+  )
+}
+
+export default ReserveInfo
